+++ src/main/java/ru/murlov/dto/ExchangeResponse.java
@@ -1,8 +1,10 @@
 package ru.murlov.dto;
 
+import java.math.BigDecimal;
+
 public record ExchangeResponse(CurrencyResponse base_currency,
                                CurrencyResponse target_currency,
-                               float rate,
-                               float amount,
-                               float convertedAmount) {
+                               BigDecimal rate,
+                               BigDecimal amount,
+                               BigDecimal convertedAmount) {
 }
